requirements.txt
# For main code
importlib_resources #==1.0.2
matplotlib #==3.1.0
Pillow #==6.0.0
scipy #==1.3.0
tabulate #==0.8.3
torch #==1.1.0
torchvision #==0.3.0
tqdm #==4.32.1

# For tests only
pytest #==4.5.0
